requirements.txt
runpod>=1.6.0
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.35.0
diffusers>=0.24.0
accelerate>=0.24.0
safetensors>=0.4.0
Pillow>=10.0.0
imageio>=2.31.0
imageio-ffmpeg>=0.4.8
xformers>=0.0.22